Implemented Guid instead of references in states
Need to implement a GUID handler for systems to use, such that the volatile objects may register themselves, and their state to the system that state requires, and from then on know that the system itself handles any state information.

+++ Venture/Assets/General/Src/SI/Kelvin.cs
@@ -17,6 +17,6 @@ namespace Assets.General.Src.SI
         { return new Kelvin() { Value = f }; }
 
         public static implicit operator Kelvin(Celsius c)
-        { return new Kelvin() { Value = c + 273.15f }; }
+        { return new Kelvin() { Value = (float)c + 273.15f }; }
     }
 }
